+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ## [0.33.0] - 2020-10-26
 
 ### Fixed
+- DropDownSelect: Removed dropdown opening logic from focus handler to click handler 
 - ToolTip: Fixed calculation of position for selecting the opening side
 - DropZone: Prohibited deleting files in disabled mode